Set chance to be a double. Change message color to light purple.

+++ src/main/java/com/minesnap/psa/Message.java
@@ -3,18 +3,18 @@ package com.minesnap.psa;
 public class Message {
     private final String name;
     private final String content;
-    private final int chance;
+    private final double chance;
     private final boolean spout;
 
     public Message(String name, String content) {
         this(name, content, 100);
     }
 
-    public Message(String name, String content, int chance) {
+    public Message(String name, String content, double chance) {
         this(name, content, 100, false);
     }
 
-    public Message(String name, String content, int chance, boolean spout) {
+    public Message(String name, String content, double chance, boolean spout) {
         this.name = name;
         this.content = content;
         this.chance = chance;
@@ -29,7 +29,7 @@ public class Message {
         return content;
     }
 
-    public int getChance() {
+    public double getChance() {
         return chance;
     }
